# backend/api/gen/go/admin/service/v1/i_file_transfer_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_file_transfer.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "go-wind-admin/api/gen/go/storage/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationFileTransferServiceDownloadFile = "/admin.service.v1.FileTransferService/DownloadFile"
const OperationFileTransferServicePostUploadFile = "/admin.service.v1.FileTransferService/PostUploadFile"
const OperationFileTransferServicePutUploadFile = "/admin.service.v1.FileTransferService/PutUploadFile"

type FileTransferServiceHTTPServer interface {
	// DownloadFile 下载文件
	DownloadFile(context.Context, *v1.DownloadFileRequest) (*v1.DownloadFileResponse, error)
	// PostUploadFile 上传文件 POST 方式
	PostUploadFile(context.Context, *v1.UploadFileRequest) (*v1.UploadFileResponse, error)
	// PutUploadFile 上传文件 PUT 方式
	PutUploadFile(context.Context, *v1.UploadFileRequest) (*v1.UploadFileResponse, error)
}

func RegisterFileTransferServiceHTTPServer(s *http.Server, srv FileTransferServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/file/download", _FileTransferService_DownloadFile0_HTTP_Handler(srv))
	r.PUT("/admin/v1/file/upload", _FileTransferService_PutUploadFile0_HTTP_Handler(srv))
	r.POST("/admin/v1/file/upload", _FileTransferService_PostUploadFile0_HTTP_Handler(srv))
}

func _FileTransferService_DownloadFile0_HTTP_Handler(srv FileTransferServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.DownloadFileRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFileTransferServiceDownloadFile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DownloadFile(ctx, req.(*v1.DownloadFileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.DownloadFileResponse)
		return ctx.Result(200, reply)
	}
}

func _FileTransferService_PutUploadFile0_HTTP_Handler(srv FileTransferServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.UploadFileRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFileTransferServicePutUploadFile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PutUploadFile(ctx, req.(*v1.UploadFileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.UploadFileResponse)
		return ctx.Result(200, reply)
	}
}

func _FileTransferService_PostUploadFile0_HTTP_Handler(srv FileTransferServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.UploadFileRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFileTransferServicePostUploadFile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PostUploadFile(ctx, req.(*v1.UploadFileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v1.UploadFileResponse)
		return ctx.Result(200, reply)
	}
}

type FileTransferServiceHTTPClient interface {
	// DownloadFile 下载文件
	DownloadFile(ctx context.Context, req *v1.DownloadFileRequest, opts ...http.CallOption) (rsp *v1.DownloadFileResponse, err error)
	// PostUploadFile 上传文件 POST 方式
	PostUploadFile(ctx context.Context, req *v1.UploadFileRequest, opts ...http.CallOption) (rsp *v1.UploadFileResponse, err error)
	// PutUploadFile 上传文件 PUT 方式
	PutUploadFile(ctx context.Context, req *v1.UploadFileRequest, opts ...http.CallOption) (rsp *v1.UploadFileResponse, err error)
}

type FileTransferServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewFileTransferServiceHTTPClient(client *http.Client) FileTransferServiceHTTPClient {
	return &FileTransferServiceHTTPClientImpl{client}
}

// DownloadFile 下载文件
func (c *FileTransferServiceHTTPClientImpl) DownloadFile(ctx context.Context, in *v1.DownloadFileRequest, opts ...http.CallOption) (*v1.DownloadFileResponse, error) {
	var out v1.DownloadFileResponse
	pattern := "/admin/v1/file/download"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFileTransferServiceDownloadFile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PostUploadFile 上传文件 POST 方式
func (c *FileTransferServiceHTTPClientImpl) PostUploadFile(ctx context.Context, in *v1.UploadFileRequest, opts ...http.CallOption) (*v1.UploadFileResponse, error) {
	var out v1.UploadFileResponse
	pattern := "/admin/v1/file/upload"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFileTransferServicePostUploadFile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PutUploadFile 上传文件 PUT 方式
func (c *FileTransferServiceHTTPClientImpl) PutUploadFile(ctx context.Context, in *v1.UploadFileRequest, opts ...http.CallOption) (*v1.UploadFileResponse, error) {
	var out v1.UploadFileResponse
	pattern := "/admin/v1/file/upload"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFileTransferServicePutUploadFile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}